Create WIndowShade
In this shader, I have used cube maps to create a perspective reflection of a city skyline. So, depending on where you are standing in relation to the window, the texture will look slightly different. However, if you get too close to the window then the texture will change to an interior cube map. This is done by calculating the distance between the user’s camera and the window and will change based on if they’re close enough. In addition to this, there is an actual camera that is used to only detect the player tag. So, when the player is visible then it will be cast onto the shader. I could have it reflect other surroundings however, in a larger scaled game, this would be bound to have performance issues.